+++ providers/basic/basic-auth_test.go
@@ -1,11 +1,12 @@
-package providers
+package basic
 
 import (
-	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 const (
